Insert no-ops into bytecode to ensure that CUDA reads are aligned
From CUDA documentation:
https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#device-memory-accesses

"Global memory instructions support reading or writing words of size
equal to 1, 2, 4, 8, or 16 bytes. Any access (via a variable or a
pointer) to data residing in global memory compiles to a single global
memory instruction if and only if the size of the data type is 1, 2, 4,
8, or 16 bytes and the data is naturally aligned (i.e., its address is a
multiple of that size)."

This commit satisfies this invariant by forcing reads to be aligned with
byte boundaries.